Updating to Phoenix 1.7.2
Rebuilt project with the Phoenix generator and replicated functionality.
Weird bugs were popping up and dependency management became too
difficult to handle. This approach allowed for a clean slate and a way
to properly handle live updates with basic channels instead of LiveView.
Still need to add channel for updating the thresholds being met in realtime
for feature parity.